+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - `renderRecent` function modified
 - `playVideo` function modified
+- `subSrc` renamed to `subtitleSrc`
+- `subType` renamed to `subtitleType`
 ## [0.9] - 2026-04-21
 ### Added
 - Video download button